requirements.txt
openai
fastdtw
pandas
numpy
scipy
google-generativeai
python-dotenv
PyYAML
matplotlib
